CLI works on check-in/check-out rentals. re-designed the cli structure.
